Add form for test fund requests (#3)
* Add form for test fund requests

Github discussions allows for [category
forms](https://docs.github.com/en/discussions/managing-discussions-for-your-community/creating-discussion-category-forms#creating-discussion-category-forms)
that can be helpful when we are requesting the same info for each
discussion point.  This commit adds an example for the test fund
requests, using the [QuickStart Guide](https://docs.entropy.xyz) as the
basis for its structure.  This would deviate from the guide in that the
Title of the discussion is set for the person, and they put their
address in the first input field of the form.

* Update .github/DISCUSSION_TEMPLATE/get-test-funds.yml

Co-authored-by: Johnny <9611008+johnnymatthews@users.noreply.github.com>

* Update .github/DISCUSSION_TEMPLATE/get-test-funds.yml

Co-authored-by: Johnny <9611008+johnnymatthews@users.noreply.github.com>

* Fix language in description

---------

Co-authored-by: Johnny <9611008+johnnymatthews@users.noreply.github.com>